save progress 27/01/2024
update javascript setting for all menu

+++ public/js/Utility/Compressor/InputPerawatan.js
@@ -54,8 +54,8 @@ jam.disabled = true;
 part.disabled = true;
 keterangan.disabled = true;
 teknisi.disabled = true;
-updateButton.disabled = true;
-deleteButton.disabled = true;
+updateButton.disabled = false;
+deleteButton.disabled = false;
 
 // Function to check if all fields are filled
 function checkAllFieldsFilled() {
@@ -97,14 +97,30 @@ inputButton.addEventListener("click", function () {
 
 // UpdateButton click
 updateButton.addEventListener("click", function () {
-    // Disable input fields and disable Update and Delete buttons
-    tanggal.disabled = false;
-    mesin.disabled = false;
-    jam.disabled = false;
-    part.disabled = false;
-    keterangan.disabled = false;
-    teknisi.disabled = false;
-    deleteButton.disabled = true;
+    var checkboxValues = $(".checkboxperawatan:checked")
+        .map(function () {
+            return this.value;
+        })
+        .get();
+
+    if (checkboxValues.length === 0) {
+        Swal.fire({
+            icon: "error",
+            title: "Tidak Ada Data Terpilih",
+            text: "Pilih satu data perawatan untuk diperbarui.",
+        });
+        deleteButton.disabled = false;
+        inputButton.disabled = false;
+    } else {
+        // Disable input fields and disable Update and Delete buttons
+        tanggal.disabled = false;
+        mesin.disabled = false;
+        jam.disabled = false;
+        part.disabled = false;
+        keterangan.disabled = false;
+        teknisi.disabled = false;
+        deleteButton.disabled = true;
+    }
 });
 
 // CancelButton click
@@ -174,8 +190,6 @@ $(document).ready(function () {
         });
     });
 });
-
-
 
 $(document).ready(function () {
     var dataTable = $("#table-perawatan").DataTable({
@@ -356,31 +370,57 @@ $(document).ready(function () {
             })
             .get();
 
-        var requestData = {
-            NoPerawatan: checkboxValues,
-        };
+        // Check if there are selected checkboxes
+        if (checkboxValues.length === 0) {
+            Swal.fire({
+                icon: "error",
+                title: "Tidak Ada Data Terpilih",
+                text: "Pilih setidaknya satu data Panel untuk dihapus.",
+            });
+            return; // Abort further processing
+        }
 
-        $.ajax({
-            url: "/delete-perawatan",
-            method: "DELETE",
-            data: requestData,
-            headers: {
-                "X-CSRF-TOKEN": csrfToken,
-            },
-            success: function (response) {
-                Swal.fire({
-                    icon: "success",
-                    title: "Terhapus!",
-                    text: "Data Berhasil Dihapus!",
-                    showConfirmButton: false,
-                    timer: "2000",
+        // Use SweetAlert for confirmation
+        Swal.fire({
+            title: "Konfirmasi",
+            text: "Anda yakin ingin menghapus data Panel yang terpilih?",
+            icon: "warning",
+            showCancelButton: true,
+            confirmButtonColor: "#3085d6",
+            cancelButtonColor: "#d33",
+            confirmButtonText: "Ya, Hapus!",
+            cancelButtonText: "Batal",
+        }).then((result) => {
+            if (result.isConfirmed) {
+                var requestData = {
+                    NoPerawatan: checkboxValues,
+                };
+                $.ajax({
+                    url: "/delete-perawatan",
+                    method: "DELETE",
+                    data: requestData,
+                    headers: {
+                        "X-CSRF-TOKEN": csrfToken,
+                    },
+                    success: function (response) {
+                        Swal.fire({
+                            icon: "success",
+                            title: "Terhapus!",
+                            text: "Data Berhasil Dihapus!",
+                            showConfirmButton: false,
+                            timer: "2000",
+                        });
+                        dataTable.ajax.reload();
+                        clearForm();
+                    },
+                    error: function (error) {
+                        console.error(
+                            "Error delete Data : ",
+                            error.responseText
+                        );
+                    },
                 });
-                dataTable.ajax.reload();
-                clearForm();
-            },
-            error: function (error) {
-                console.error("Error delete Data : ", error.responseText);
-            },
+            }
         });
     });
 });
